requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
google-generativeai
google-cloud-speech
google-cloud-storage
pydantic
pydantic-settings
librosa
praat-parselmouth
numpy
scipy
imageio
imageio-ffmpeg
llama-cpp-python
grpcio
protobuf
huggingface-hub